Add option to get vol & tracer bulk values at once
So far, it has been necessary to calculate the bulk values for volume
transport and tracer transport separately.  While it was possible to
provide the indices of the bulk values of the first calculation to the
second function call, reducing the number of computations, this method
did not allow to specify which bulk values correspond to inflow and
which to outflow.  This can be unclear when the tracer changes sign
(e.g., temperature in high latitudes), leading to a wrong classification
of in- and outflow bulk values.

This commit implements a new option to compute the bulk values for
volume transport and tracer transport at once.  This reduces the
required computations and has the benefit that it is always clear which
bulk values correspond to in- and which to outflow.  Furthermore, it is
more user-friendly, since only a single Dataset is returned by the one
function call, instead of two separate Datasets from two function calls.

The new option is optional, so it should not break any existing code.

The function `calc_bulk_values` contained `**kwargs` in its header,
but `kwargs` was not used anywhere in the function body, so it is
removed with this commit to avoid possible confusion.